+++ extensions/finder/src/main/java/org/patternfly/extension/finder/FinderItem.java
@@ -33,6 +33,7 @@ import org.patternfly.core.Aria;
 import org.patternfly.core.ComponentContext;
 import org.patternfly.core.Dataset;
 import org.patternfly.style.Classes;
+
 import elemental2.dom.Element;
 import elemental2.dom.Event;
 import elemental2.dom.HTMLElement;